impl(service): findByUserName implement to EventsRepository

+++ selfouettie-server/src/main/scala/net/yoshinorin/selfouettie/types/db/Manipulation.scala
@@ -12,4 +12,4 @@ object Limit {
   }
 }
 
-case class FromTo(from: Option[Int] = None, to: Option[Int])
+case class FromTo(from: Option[Long] = None, to: Option[Long])
